requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
PyJWT>=2.8.0
cryptography>=41.0.0
bcrypt>=4.1.0
